client/usb_deployment/client/requirements.txt
# Python Tenjo Client Dependencies

# Core libraries
requests>=2.31.0
websocket-client>=1.6.0
psutil>=5.9.0
setproctitle>=1.3.3

# Screen capture
mss>=9.0.0
Pillow>=10.0.0

# Windows-specific (will be ignored on other platforms)
pygetwindow>=0.0.9; platform_system=="Windows"
pywin32>=306; platform_system=="Windows"
wmi>=1.5.1; platform_system=="Windows"

# macOS-specific (for native screen recording permissions and functionality)
pyobjc-core>=10.0; platform_system=="Darwin"
pyobjc-framework-Quartz>=10.0; platform_system=="Darwin"
pyobjc-framework-Cocoa>=10.0; platform_system=="Darwin"
